# api/openapi.yaml
openapi: 3.0.3
info:
  title: Spillway admin API
  version: 0.4.0
  description: |
    The API behind the dashboard. Its only caller is the dashboard's server-side layer, which signs users in
    with POST /admin/login and forwards the returned token as `Authorization: Bearer <token>`. The browser
    never sees the token. TypeScript types for web/ are generated from this file, and a Go test checks that
    the real responses match it.

    Each operation says which roles may call it with `x-access`: `public`, `viewer` (any signed-in user)
    or `admin`. The Go router registers the same access level, and a test fails if they disagree.
paths:
...
  /admin/tools:
    get:
      operationId: toolsList
      summary: The tool registry
      description: HTTP tools and MCP servers. Auth header values are never returned, only their names.
      x-access: viewer
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Registered tools
          content:
            application/json:
              schema:
                type: object
                required: [tools]
                properties:
                  tools: { type: array, items: { $ref: "#/components/schemas/Tool" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: toolCreate
      summary: Register a tool or an MCP server
      x-access: admin
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ToolInput" }
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tool" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: A tool with that name exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /admin/tools/{id}:
    put:
      operationId: toolUpdate
      summary: Change a tool
      description: Fields left out are kept. Headers, when sent, replace the stored ones.
      x-access: admin
      security: [{ bearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ToolUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tool" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: toolDelete
      summary: Remove a tool
      x-access: admin
      security: [{ bearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Removed }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/tools/{id}/discover:
    post:
      operationId: toolDiscover
      summary: Fetch an MCP server's tools
      description: Calls tools/list on the server and stores the answer. Its tools are then usable in a run as `server.tool`.
      x-access: admin
      security: [{ bearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The server with its discovered tools
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tool" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502":
          description: The MCP server could not be reached or answered badly
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
components:
...
  responses:
...
    BadRequest:
      description: The request is not valid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No token, a bad token, or an expired one
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The signed-in user is a viewer and this needs an admin
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...
    Tool:
      type: object
      required: [id, name, kind, endpoint, description, header_names, timeout_ms, requires_approval, discovered_at, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        kind: { type: string, enum: [http, mcp] }
        endpoint: { type: string }
        description: { type: string }
        input_schema: { type: object, additionalProperties: true, description: "http tools: the JSON Schema of the arguments" }
        mcp_tools:
          type: array
          description: "mcp servers: what the last discover found. Names are `server.tool`."
          items:
            type: object
            required: [name, description]
            properties:
              name: { type: string }
              description: { type: string }
              input_schema: { type: object, additionalProperties: true }
        header_names: { type: array, items: { type: string } }
        timeout_ms: { type: integer, format: int64 }
        requires_approval: { type: boolean }
        discovered_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    ToolInput:
      type: object
      additionalProperties: false
      required: [name, endpoint]
      properties:
        name: { type: string }
        kind: { type: string, enum: [http, mcp], default: http }
        endpoint: { type: string }
        description: { type: string }
        headers: { type: object, additionalProperties: { type: string } }
        input_schema: { type: object, additionalProperties: true }
        timeout_ms: { type: integer, format: int64 }
        requires_approval: { type: boolean }
    ToolUpdate:
      type: object
      additionalProperties: false
      properties:
        endpoint: { type: string }
        description: { type: string }
        headers: { type: object, additionalProperties: { type: string } }
        input_schema: { type: object, additionalProperties: true }
        timeout_ms: { type: integer, format: int64 }
        requires_approval: { type: boolean }
...